requirements.txt
discord.py
python-dotenv
Pillow
aiosqlite
aiohttp
asyncpg
